Changed folder users with timestamp users

+++ super/includes/delete-super.inc.php
@@ -8,30 +8,20 @@ if(($_SESSION['u_sid']=="AmiXchange") && ($_GET['ur']=="/AmiXchange/super/home-s
    $resultcheck=mysqli_num_rows($result);
    if($resultcheck>0)
    {
-       while($row=mysqli_fetch_assoc($result))
-       {
-           
-          
-               $file="../../assets/Users/".$row['user_dp'];
-               $ads="../../assets/Products/".$_GET['id'];
-               unlink($file);
-              /* 
-               * php delete function that deals with directories recursively
-               */
-             function delete_files($target) {
-               if(is_dir($target)){
-                  $files = glob( $target . '*', GLOB_MARK ); //GLOB_MARK adds a slash to directories returned
-
-                  foreach( $files as $file ){
-                    delete_files( $file );      
-                   }
-                   } elseif(is_file($target)) {
-                                       @chmod($target,0644);
-                                       unlink( $target );  
-                                             }
-                                            }         
-               delete_files($ads);
-               rmdir($ads);
+    $row=mysqli_fetch_assoc($result); 
+    $file="../../assets/Users/".$row['user_dp'];
+    $ads="../../assets/Products/";
+    unlink($file);
+    $sql1="select Product_Pic from advertisements where user_uid='".$_GET['id']."';";
+    $result1=mysqli_query($conn,$sql1);
+    $resultcheck1=mysqli_num_rows($result1);
+    if($resultcheck1>0)
+    {
+     while($row1=mysqli_fetch_assoc($result))
+     {
+      $file1="../../assets/Products/".$row1['Product_Pic'];
+      unlink($file);
+     }                       
                $del_sql1="delete from users where user_uid='".$_GET['id']."'";
                $del_sql2="delete from verification where user_uid='".$_GET['id']."'";
                $del_sql3="delete from advertisements where user_uid='".$_GET['id']."'";
@@ -42,5 +32,5 @@ if(($_SESSION['u_sid']=="AmiXchange") && ($_GET['ur']=="/AmiXchange/super/home-s
                    
        }
    }
-}
+  }
 ?>
